perf: opt2-opt14 - BVH flat-tree fix, SIMD 4-wide AABB, shadow/shading O3 attrs, float upscale, exp2, clamp_d inline

+++ cercle-5/miniRT/includes/functions/helpers/random.h
@@ -21,8 +21,8 @@
 ** rt_seed_mix  : XORs cached base with frame+id (cheap, call per sample).
 ** rt_seed_point: convenience wrapper (pos + mix in one call).
 */
-uint32_t	rt_seed_pos(t_vec3 p);
-uint32_t	rt_seed_mix(uint32_t base, int frame, int id);
-uint32_t	rt_seed_point(t_vec3 p, int frame, int id);
+uint64_t	rt_seed_pos(t_vec3 p);
+uint64_t	rt_seed_mix(uint64_t base, int frame, int id);
+uint64_t	rt_seed_point(t_vec3 p, int frame, int id);
 
 #endif
